findInText() had a bug and GOF formula reverts
Getting down to it now. Had written findInText() before later using assumption that nars could preserve their "found" state while clearing and re-using their ifound[]. This was mis-handled in the  return value of the function. Meanwhile, the GOF formula has been giving me trouble because the "it" in "it was very quiet" produced jfound[0]=0 and this gives a GOF of 1.0 in the last control segment of "Although our room was near the elevator, it was very quiet". That is because r=1 and f=1 in the r/f part of the GOF formula. Take your pick: add a penalty for f being smaller than the sub range of tokens where it is computed OR get "it" out of the picture. I did the latter for now, and "it was" becomes an entry in the kEXPERIENCE variable.