another deep bug
Polarity was not getting relayed back up the nar's tree. Now findInText( ) uses the child.polarity when it is relevant.
Also: gonna be futzing with polarity combination algorithms for a while. This time, for readTextAsAttribute() I added code to handle incomplete matching taking its polarity value from a "thing" when there is no "value".