Segment now allows multiple VARS possible for one itok
findInText2( ) has been returning with the first VAR found in tree to match the text. This meant being very strict about different VARS sharing (or not) the same keywords. Now findInText2( ) returns a list of all vars in tree found to match a given itok index. This returns to Narwhal to V1 behavior. The advantage is that the client may need to deal with words like "next door" that both express a proximity relation AND imply being in a hotel or a hotel room. The client should be able to use both.