Added support for NAR.lastConst and 4 accessors
Kind of a nice system. All the ReadSegment( ) subroutines now set the lastConst to a ':' separated list of most recently read VAR knames[0]. This composite string can be split and used to return NAR.Thing(), NAR.Action(), NAR.Relation(), and NAR.Value(). This closes the loop on data storage - it is stored in the NAR but it is volatile and can change on the way through the text.